- moving time and getDay methods to helpers - slight refactor
